rust/tx: optimize transaction iterator
Ticket: 7089

App-layer protocols have transaction iterators.
These transaction iterators are implemented by the app-layer,
based on the structure used for the list of transactions :
it may be a C linked list with pointers, a rust VecDeque, etc...
These tx iterators are used in different places: detection, logging,
and cleanup.
Cleanup removes the transactions from their list structure.

So, we face the classical problem coming of removing elements of a list
while iterating over it.

The current solution, for rust, implemented by a trait,
is to propose the "next" element (variable state)
as the same that was returned.

So, in most cases, next call of the iterator, restarts from
the previously returned element, and makes a dummy iteration
of the loop in get_transaction_iterator to move forward.
In the case where the tx got removed,
there is no dummy iteration and it returns quickly.

This commit optimizes the first case which happens more often,
and remain correct in the more complex case with tx removal...

The solution is to not have removal of transactions while iterating
over their list, as it is a classic treacherous pattern.
So, the solution is to have a first loop iterating over the transactions
and retaining the tx ids to remove.
And then iterate over these tx ids to remove the transactions.